rename package from it to demo
